Restructure mongoDB client
