+++ mongodb/src/mongodb/errors.bal
@@ -14,22 +14,37 @@
 // specific language governing permissions and limitations
 // under the License.
 
-# Holds the details of an MongoDB error
+# Holds the properties of an Application error
 #
 # + message - Specific error message for the error
 # + cause - Cause of the error
-# + detail - Error detail
-public type ErrorDetail record {
+public type ApplicationErrorDetail record {
     string message;
     error cause?;
-    string detail?;
+};
+
+# Holds the properties of a Database error
+#
+# + message - Specific error message for the error
+# + mongoDBExceptionType - Type of the mongodb exception returned
+# + cause - Cause of the error
+public type DatabaseErrorDetail record {
+    string message;
+    string mongoDBExceptionType;
+    error cause?;
 };
 
 // Error reasons
 public const DATABASE_ERROR_REASON = "{ballerina/mongodb}DatabaseError";
+
+# Represents an error caused by an issue related to database accessibility, erroneous queries, constraint violations,
+# database resource clean-up, and other similar scenarios.
+public type DatabaseError error<DATABASE_ERROR_REASON, DatabaseErrorDetail>;
+
 public const APPLICATION_ERROR_REASON = "{ballerina/mongodb}ApplicationError";
 
-public type DatabaseError error<DATABASE_ERROR_REASON, ErrorDetail>;
-public type ApplicationError error<APPLICATION_ERROR_REASON, ErrorDetail>;
+# Represents an error originating from application-level causes.
+public type ApplicationError error<APPLICATION_ERROR_REASON, ApplicationErrorDetail>;
 
-public type ConnectorError DatabaseError|ApplicationError;
+# Represents a database or application level error returned from MongoDB client remote functions.
+public type Error DatabaseError|ApplicationError;
